+++ projects/minecraft/latest/forge/src/main/java/com/breakinblocks/plonk/common/command/IPlonkCommand.java
@@ -18,7 +18,7 @@ public interface IPlonkCommand {
     }
 
     default int sendUsage(CommandSourceStack source, int exitCode) {
-        source.sendSuccess(Component.literal(getUsage(source)), true);
+        source.sendSuccess(() -> Component.literal(getUsage(source)), true);
         return exitCode;
     }