+++ projects/minecraft/latest/forge/src/main/java/com/breakinblocks/plonk/common/registry/RegistryMaterials.java
@@ -1,8 +1,0 @@
-package com.breakinblocks.plonk.common.registry;
-
-import net.minecraft.world.level.material.Material;
-import net.minecraft.world.level.material.MaterialColor;
-
-public class RegistryMaterials {
-    public static final Material placed_items = (new Material.Builder(MaterialColor.NONE)).notSolidBlocking().build();
-}