+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEXT
 
 - Ported to Minecraft 1.19.4 for Forge.
+- Ported to Minecraft 1.20.1 for Forge.
 
 ## [10.0.4] - 2023-04-08